Prototyp/backend/requirements-prod.txt
# Production requirements for Koyeb deployment
# Optimized for minimal memory footprint and cloud compatibility

# Core FastAPI stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
aiofiles==23.2.1

# HTTP client for API calls
requests>=2.31.0

# Medical data handling
fhir.resources>=7.0.0

# Vector database (CPU-only for cloud)
faiss-cpu==1.7.4
numpy>=1.24.0

# Database (PostgreSQL for cloud)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# CORS and security
python-jose[cryptography]>=3.3.0

# Optional: Monitoring and logging
structlog>=23.1.0

# Memory optimization
pympler>=0.9  # Memory profiling for optimization